requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.0.9
django-admin-autocomplete-filter==0.6.1
django-admin-filter==0.1
django-admin-rangefilter==0.6.3
django-author==1.0.2
django-braces==1.13.0
django-cors-headers==3.5.0
django-csp==3.7
django-filter==2.4.0
django-jet==1.0.8
django-nucleus==0.1.2
django-object-actions==3.0.1
gunicorn==20.0.4
jdatetime==3.6.2
jsonfield==3.1.0
persiantools==1.5.1
psycopg2==2.8.6
pytz==2020.4
setuptools-git==1.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
